I’ve finished the weighted product support now. Products can be marked as weighted with a unit (like kg), and that data is stored in the database. I updated the product model, DAO, admin stock controller, cart logic, and order saving so weighted items display properly in the cart (e.g. quantity as kg) and calculate subtotal/total correctly based on quantity × unit price. It’s all working end-to-end now, including checkout.

+++ src/main/java/com/till/dao/SalesDAO.java
@@ -49,7 +49,7 @@ public class SalesDAO {
                     itemStmt.setLong(1, orderId);
                     itemStmt.setString(2, item.getProduct().getId());
                     itemStmt.setString(3, item.getProduct().getName());
-                    itemStmt.setInt(4, item.getQuantity());
+                    itemStmt.setDouble(4, item.getQuantity());
                     itemStmt.setDouble(5, item.getProduct().getPrice());
                     itemStmt.setDouble(6, item.getSubtotal());
                     itemStmt.addBatch();
